# packages/server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: VaultGuard API
  version: 0.1.3
  description: |
    HTTP contract that a VaultGuard-compatible backend must implement.

    This document is the **client-facing contract** extracted from the
    Obsidian plugin's API client (`src/api/client.ts`) and key-lease
    manager (`src/crypto/key-manager.ts`). It enumerates every endpoint
    the plugin actually calls, the request/response shapes it expects,
    and the auth scheme it uses.

    **For self-hosters.** VaultGuard Community Edition ships an AWS Lambda +
    API Gateway + Terraform reference server, distributed under the
    Sustainable Use License. Any compatible backend that satisfies this
    contract will work with the unmodified plugin.

    ## Architectural rules embodied in this contract

    - **Vault scoping.** Every file and permission endpoint lives under
      `/vaults/{vaultId}/...`. There is no root-level `/files/*` or
      `/permissions/*`. See `docs/VAULTS.md`.
    - **Authentication.** Bearer-style auth using a Cognito-issued
      (or compatible OIDC) ID token, passed as the raw token in the
      `Authorization` header. **No `Bearer ` prefix** for the main API
      surface (see `getAuthHeaders` in `client.ts`). The key-lease
      endpoints under `/auth/*` accept the token with a `Bearer ` prefix
      (see `key-manager.ts`); compliant servers should accept either.
    - **Error shape.** Failed responses return JSON matching
      `#/components/schemas/ApiError`.
...

...

...
servers:
...
security:
  - cognitoIdToken: []
paths:
...
  /vaults/{vaultId}/shares:
    parameters:
      - $ref: '#/components/parameters/VaultId'
    post:
      tags: [Shares]
      summary: Create a share link for a file in this vault
      description: |
        Mints an opaque pointer to a vault-relative path. Returns the public
        https URL to copy into Slack/email. Caller must (a) be a vault
        member, and (b) have `read` permission on the path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [relPath]
              properties:
                relPath:
                  type: string
                  description: Vault-relative path of the file (no leading slash required).
                  example: Notes/Welcome.md
                expiresAt:
                  type: string
                  format: date-time
                  description: |
                    Optional ISO-8601 timestamp after which the link is
                    rejected. Must be in the future and within 1 year.
      responses:
        '201':
          description: Share created
          content:
            application/json:
              schema:
                type: object
                properties:
                  share: { $ref: '#/components/schemas/ShareRecord' }
                  url:
                    type: string
                    format: uri
                    example: https://share.example.com/s/abcDEF123?v=vault-uuid
        '400':
          description: Invalid relPath, malformed expiresAt, or expiry out of range
        '403':
          description: Caller has no read permission for this path
    get:
      tags: [Shares]
      summary: List active share links for this vault
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        '200':
          description: Active shares (expired entries are filtered out)
          content:
            application/json:
              schema:
                type: object
                properties:
                  shares:
                    type: array
                    items: { $ref: '#/components/schemas/ShareRecord' }
                  count: { type: integer }
  /vaults/{vaultId}/shares/{shareId}:
    parameters:
      - $ref: '#/components/parameters/VaultId'
      - in: path
        name: shareId
        required: true
        schema: { type: string }
    get:
      tags: [Shares]
      summary: Resolve a share token to its (vault, path) target
      description: |
        Used by the plugin's `obsidian://vaultguard-share` URI handler.
        Caller must be a vault member AND have `read` permission on the
        underlying path; otherwise the endpoint returns 404 (not 403) so
        a token can't be probed for existence by an unauthorised member.
      responses:
        '200':
          description: Resolution success
          content:
            application/json:
              schema:
                type: object
                properties:
                  shareId: { type: string }
                  vaultId: { type: string }
                  vaultName: { type: string }
                  vaultSlug: { type: string }
                  relPath: { type: string }
                  createdBy: { type: string }
                  createdAt: { type: string, format: date-time }
                  expiresAt: { type: [string, 'null'], format: date-time }
        '404':
          description: Token not found, belongs to another vault, or recipient lacks read
        '410':
          description: Share link has expired
    delete:
      tags: [Shares]
      summary: Revoke a share link
      description: |
        Creator can always revoke their own share. Vault admins can revoke
        any share in the vault.
      responses:
        '200':
          description: Revocation success
        '404':
          description: Share not found
        '403':
          description: Caller is not the creator and not a vault admin

  # ─── Users (org-level) ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    cognitoIdToken:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Cognito ID token (or compatible OIDC ID token) sent as the raw token
        in the `Authorization` header — **no `Bearer ` prefix** for the main
        API surface. The `/auth/key-lease`, `/auth/refresh`, `/auth/heartbeat`,
        and `/auth/wrapped-key` endpoints are called by the key-manager
        module which sends `Bearer <token>`; compliant servers should accept
        either form.
  parameters:
    VaultId:
      in: path
      name: vaultId
      required: true
      schema: { type: string }
      description: Server-side vault ID. Throws client-side if unbound.
...
  schemas:
...
    ShareRecord:
      type: object
      required: [shareId, vaultId, relPath, createdAt, createdBy]
      properties:
        shareId:
          type: string
          description: Random URL-safe token. PK in the Shares table.
        vaultId: { type: string }
        relPath:
          type: string
          description: Vault-relative path of the shared file at mint time.
        createdAt: { type: string, format: date-time }
        createdBy:
          type: string
          description: User ID of the creator. They (and vault admins) can revoke.
        expiresAt:
          type: string
          format: date-time
          description: Optional ISO timestamp after which the link is rejected.
          nullable: true
        url:
          type: string
          format: uri
          description: Full https URL — copy into Slack/email.
...